Change YouTube parser to support Wordpress.com syntax.
According to <http://en.support.wordpress.com/videos/youtube/>, the
preferred syntax for YouTube shortcodes (unlike those for other video
websites) is to include the full URL. Including an "id" parameter is not
supported on wordpress.com.

Note that this parser, like the original django_shortcodes one, does not
support the rel and showsearch parameters that wordpress.com does.